[IMP] l10n_in*/saas~16.1: remove multi GSTIN fields and views
Multi-unit is created if there is a different branch with the same GSTIN(vat) so
now we decided to create a multi-company instead of this.

because functionally we discuss with Indian CA and we find out that it's good
practice if we create a different company.

The main problem is one company and the same Bank account but we also find a
functional solution using a transfer balance to another company.

so we remove this so our other integrations do not have complexity
like E-invoice, E-waybill, GST E-filing

related PR: https://github.com/odoo/odoo/pull/107234

TaskID - 2723109

closes odoo/upgrade#4107

Related: odoo/odoo#107234
Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>