[FIX] base/16.0: convert to BS5 no update views
The bootstrap 5 migration script does not convert the views having a
xml_id, because they will be updated by the content of the xml file
containing it after the migration. But if the view is in no update,
there won't be such update, and you could keep a some views that have
not been converted by the migration scipt.

So instead of just checking that the view has an xml_id, we are also
checking that the view is not in no update to discard it from the views
to migrate