[FIX] util/orm: don't flush always
The flush is not needed always. It is undesirable in the crawler. This
extra flush was added in d13d617f mainly to support the fixing of views
in `pre-models-ir_ui_view.py` because we edit the archs.

If we always flush we may get an error due to failed computes, and that
fails the whole upgrade.

upg-424429

Part-of: odoo/upgrade#4148