[FIX] hr_payroll/saas~15.4: imply new group
Since a group has been created in the following PR: odoo/enterprise/pull/23669
The Payroll Officer is suppose to imply the group of Payroll Manager
which has less rights but is used in many views.

As the group records are in no update, after the migration, the payroll
officer doesn't have the access of the group manager, and the groups are
not grouped in a selection field in the user access rights page.

With this script we imply the group manager in the officer one.