[FIX] hr_contract/saas~15.4: imply new group
Since a new group has been created in the following PR: odoo/odoo#83191
The Contract Administrator is suppose to imply the group of ContractEmployee
Manager which has less rights but is used in many views.

As the group records are in no update, after the migration, the contract
administrator doesn't have the access of the group employee manager, and the
groups are not grouped in a selection field in the user access rights page.

With this script we imply the group emploee manager in the administrator one.