[FIX] account/saas~13.2.1.1: make deprecated regexps optional
In the current implementation, if the `sequence_override_regex` on a
journal makes use of non-deprecated fields like `year` or `month`, they
will be wrapped in an optional capture group. In the case of deprecated
fields like `day`, they are replaced with mandatory substrings like
`\d{2}`. This means that the regular expression will not be able to
match the names `''` and `'/'` when creating new invoices in certain
workflows, like from a sales order, resulting in a traceback. Odoo
assumes the regular expression should match those strings, hence
wrapping them in an optional non-capturing group here. The non-capturing
part is important to avoid interfering with existing Odoo code that
makes assumptions about the capture groups present.

upg-423396
opw-3101804

closes odoo/upgrade#4169

Signed-off-by: Christophe Simonis <chs@odoo.com>