[FIX] base/0.0.0: update custom fields
By default when we load a module `_auto_init` does not update custom
field columns:
https://github.com/odoo/odoo/blob/fae1f0e363587ac743cac2a7102528f1db4ec83b/odoo/models.py#L2600

In 16.0 since we changed translated fields to be jsonb instead of text
columns. We adapt the columns during module loading, more precisely when
the column of the fields is updated.
```
-> registry.init_models(cr, model_names, {'module': package.name}, new_install)
  /home/odoo/src/odoo/16.0/odoo/modules/registry.py(453)init_models()
-> model._auto_init()
  /home/odoo/src/odoo/16.0/odoo/models.py(2602)_auto_init()
-> new = field.update_db(self, columns)
  /home/odoo/src/odoo/16.0/odoo/fields.py(984)update_db()
-> self.update_db_column(model, column)
  /home/odoo/src/odoo/16.0/odoo/fields.py(1861)update_db_column()
-> super().update_db_column(model, column)
  /home/odoo/src/odoo/16.0/odoo/fields.py(1020)update_db_column()
-> self._convert_db_column(model, column)
  /home/odoo/src/odoo/16.0/odoo/fields.py(1636)_convert_db_column()
-> sql.convert_column_translatable(model._cr, model._table, self.name, self.column_type[1])
  /home/odoo/src/odoo/16.0/odoo/tools/sql.py(137)convert_column_translatable()
-> _convert_column(cr, tablename, columnname, columntype, using)
  /home/odoo/src/odoo/16.0/odoo/tools/sql.py(147)_convert_column()
```

Issue: if we have a stored related manual field pointing to a standard
field, then we may get an error during module load if such field gets
updated.

Example:
```
(Pdb) _logger.error("bad query: %s\nERROR: %s", tools.ustr(self._obj.query or query), e)
2023-01-05 12:34:38,075 22 ERROR 421543_afu_15.0_16.0 odoo.sql_db: bad query: UPDATE "mrp_eco_stage" SET "approval_roles" = '', "is_blocking" = false, "write_date" = '2023-01-05T12:28:26.510532'::timestamp, "write_uid" = 1, "x_studio_type" = '{}' || COALESCE("x_studio_type", '{}'::jsonb) || '{"en_US": "BOM Updates"}' WHERE id IN (37)
ERROR: COALESCE types character varying and jsonb cannot be matched
LINE 1: ..._studio_type" = '{}' || COALESCE("x_studio_type", '{}'::json...
```
```
421543_afu_15.0_16.0=> select id,model,name,translate,store,related,ttype,field_description from ir_model_fields where (name in ('x_studio_type', 'type_ids') and model='mrp.eco.stage') or (name='name' and model='mrp.eco.type')
+-------+---------------+---------------+-----------+-------+---------------+-----------+--------------------+
| id    | model         | name          | translate | store | related       | ttype     | field_description  |
|-------+---------------+---------------+-----------+-------+---------------+-----------+--------------------|
| 9834  | mrp.eco.type  | name          | True      | True  | <null>        | char      | {"en_US": "Name"}  |
| 14177 | mrp.eco.stage | x_studio_type | True      | True  | type_ids.name | char      | {"en_US": "Type"}  |
| 9895  | mrp.eco.stage | type_ids      | False     | True  | <null>        | many2many | {"en_US": "Types"} |
+-------+---------------+---------------+-----------+-------+---------------+-----------+--------------------+
```
Note that the related could come from a model that is loaded _after_ the
pointed to standard field/model was loaded. Thus the target field was
already updated but at the time the custom field was loaded it was never
updated. Then upon update of the value of the custom field and error is
raised.

The solution we propose here is to always force the update of custom
fields during the upgrade.

upg-421543

closes odoo/upgrade#4179

Signed-off-by: Christophe Simonis <chs@odoo.com>