[IMP] {point_of_sale,pos_restaurant}/saas~16.2.1.0: m2o to m2m
The PR of the community add a cross-order feature for the PoS application.
The PoS config have now access to past orders of the same company
and to open orders of PoS config that are linked to it
through the "share_order" variable for retails
and through the restaurant.floor in restaurant which became a M2M field.

Odoo - odoo/odoo#109216
Enterprise - odoo/enterprise#37955

closes odoo/upgrade#4393

Signed-off-by: Christophe Simonis <chs@odoo.com>