[IMP] point_of_sale/saas~16.1.1.0: Add shipping_date, remove to_ship
The PR of the community version adds a date picker in the pos config
so that the users can choose a date when they enable the options "ship later"
for an order. It also display the delivery date on the receipt if there is one.
The field to_ship is thus not needed anymore since the new field shipping_date
has all the necessary information.

See odoo/odoo#107115

closes odoo/upgrade#4200

Signed-off-by: Christophe Simonis <chs@odoo.com>