[FIX] account_asset/16.0: fix prorata date
When the fiscal year last month and day are before the current month and
day the computation of the previous period most be done differently. For
clarity, lets consider we are at 2020-10-01 and the following two cases:
1. The end of period is on Dec 31st. In this case we know the start of
   the current period was on 2020-01-01. We can say that the date is
   then (current_year-1, Dec, 31) + 1 day and that's correct.
2. The end of period is in Mar 31 (like India, for example). In this
   case the current period started on 2020-04-01. Which is
   (current_year, Mar, 31) + 1

There is an important difference between those cases: we cannot blindly
subtract a year to the current year. That's the reason why the upgrade
of l10n_in_reports is broken before this fix.

closes odoo/upgrade#4206

Signed-off-by: Christophe Simonis <chs@odoo.com>