[FIX] purchase/16.0: do not remove price diff accounts
Community PR odoo/odoo#108513

We removed the price difference account fields on products and category
following a functional mistake (explained in community PR).
So we have to reintroduce them and keep them in the fix module
during migration

for 16.0. In 16.1 they will be put and migrate back to purchase_stock.

Part-of: odoo/upgrade#4315