[FIX] stock/16.0: remove replenishment flag from locations with wrong wh
The warehouse id on stock locations is now stored, and the replenishment
orders no longer get it from the stock quantity report.
Which means now that the warehouse id must be correctly set on replenishment
locations, which are the warehouses' lot_stock_id. The warehouses
must have the correct view location as a location of usage 'view'
and as a parent location to its lot stock location.

When the replenishment location has a wrong warehouse id, the replenishment
flag is removed and a note is added to the report to let the user fix
their location hierarchies and warehouses

https://github.com/odoo/odoo/pull/97109

closes odoo/upgrade#4286

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>