[IMP] stock/saas~16.2: delete packaging without product
The `product_id` is required on `product.packaging` but because of a
missing context, it was possible, in v16.1 and before, to create a
product's packaging without product from the picking's form view.

As this field is readonly once the packaging is saved and as packaging
without `product_id` can't be used anyway, they are useless and deleting
them seems to be the best solution.

task-3076044
Community PR: odoo/odoo#109511

Part-of: odoo/upgrade#4252