[FIX] account/16.0: set payment sequence to False
If existing journals, with existing payments and misc entries suddenly
start having different sequences the `sequence.mixin` will try to
attribute the same sequence to 2 different documents.
For instance, if we have this before the upgrade
* BNK/0001 (payment)
* BNK/0002 (misc entry)

Then if we set the payment sequence to True, and we create a new
payment, the next sequence will be
* BNK/0002 (payment)
Which is forbidden because the sequence must be unique for the whole
journal.

closes odoo/upgrade#4292

Signed-off-by: Christophe Simonis <chs@odoo.com>