[FIX] base/saas~14.3.1.3: Avoid assets infinite recursion
When trying to calculate the correct sequence number for assets, we first
check their target's (if available) sequence and do so recursively, until
there's no target within the same bundle.
If an asset's target and path are the same, then the asset will always match
itself, resulting in infinite recursion and eventually failing when the
sequence number grows out of bounds.

This fix ensures that the asset cannot match itself.

closes odoo/upgrade#4327

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>