[FIX] website_enterprise/16.0.1.0: fix nested with
The inner with() used to open cowed views within a loop is overwriting
variables in the outer with() (used to open the original, standard view).
Specifically, arch_etree would be set to the arch of the latest cow
view, thus when it returns to the generator of the outer with(), it
creates new_arch based on the original arch of that cowed view and not
the arch of the original, standard view.

closes odoo/upgrade#4303

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>