[IMP] account/saas~16.2: remove real_amount field
Remove the `real_amount` field on `account.tax` which is now useless.

To be able to compute the taxes offline, the POS JS code mimicked the
`compute_all` function in `account_tax.py`. At some point, the POS added
an extra field on `account.tax` (see:
https://github.com/odoo/odoo/commit/8fb53c53c3128e8cea7ef20b9ab4946ac2f9b7d9):
`real_amount` which is defined as:
real_amount = tax.amount * sum(tax.repartition_line.factor)
Then, when fetching the data from the server to the client, the
`real_amount` was passed in place of the `amount`.

But then, the `amount` field in the JS is no longer the same as in the
python code and it is very error prone. In addition, it's easy to get
rid of this extra field by using the sum of the factors from the
repartition lines.

Now, we pass the sum(tax.repartition_lines.factor_percent)/100, which is
equivalent to sum(tax.repartition_line.factor) to the client, and use it
in the `compute_all` as done in the python function, and leave
`tax.amount` untouched.

Both `compute_all` functions now operate using the same `tax.amount` and
we get rid of the `real_amount` field.

see https://github.com/odoo/odoo/pull/112063.

closes odoo/upgrade#4307

Related: odoo/odoo#112063
Signed-off-by: Christophe Simonis <chs@odoo.com>