[IMP] l10n_{dz,in,ua,ma},account/saas~16.2: invoices-amount in letters
The amount in letters on invoices was previously a feature specific to
some localizations such as l10n_dz. It is now a generic feature and the
record needs to be removed. This commit handles the migration of the
move of the feature from some of the localization to the account module.
Also adds the field 'display_invoice_amount_total_words'

See odoo/odoo#107714
See odoo/enterprise#35959
task : 3097097

closes odoo/upgrade#4326

Signed-off-by: Christophe Simonis <chs@odoo.com>