[IMP] base/saas~16.2: remove test_web_grid module
This module has been removed because all python methods only used for
the grid view will be removed in the related PR to use the basic ORM
methods (`read_group` instead of `read_grip` or `read_grid_grouped` for
instance). Also, the columns in the header of the grid view are no
longer generated in python, they are generated in JS based as it is the
case for the gantt view for instance.

Community PR: odoo/odoo#108605
Enterprise PR: odoo/enterprise#35217

task-2702636

closes odoo/upgrade#4410

Signed-off-by: Christophe Simonis <chs@odoo.com>