[IMP] mass_mailing/saas~16.2: remove A/B percentage limit
The limit for 100% total percentage across mailings of the same campaign
was removed.

In the process we also add a field to identify which mailing is the winner
as the previous assumption that only the winner could be 100% is now broken.

We can store that value on the campaign for each mailing to have access
to it and as a consequence `ab_testing_completed`
can be inferred from the state of
that field.

Remove now unused `ab_testing_total_pc`
computed field.
Populate `ab_testing_winner_mailing_id`
so older mailings are consistent.

Community PR: odoo/odoo#105639

task-3008627

closes odoo/upgrade#4257

Signed-off-by: Christophe Simonis <chs@odoo.com>