[IMP] helpdesk/saas~16.1: remove ir.rule introduced in the same version
The ir.rule has been introduced in 16.1 but it was a bad idea and
has been removed in the related PR in the same version.
The reason is an access error could be raised because of that rule when
the user tries to create a helpdesk team because a helpdesk stage will
be fetched as default stage for the new helpdesk team, however, that
stage could be linked to a helpdesk team in another company and so the
rule will not be satisfied.

Enterprise PR: odoo/enterprise#36868

closes odoo/upgrade#4348

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>