[FIX] hr_expense/16.0: not possible to update invoice_date
Before v16, posted expenses are of move_type 'entry'
After v16, of type 'in_receipt'.
But internally, we treat them as 'in_invoice' (bills).

So if a user created and posted an expense in v15, migrate to v16,
and want to update the move. He will get 'The Bill/Refund date
is required to validate this document.'

This commits set an invoice_date for those moves.

opw-3171640

closes odoo/upgrade#4341

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>