[FIX] account/16.0.1.2: migrate analytic defaults
Analytic Defaults have been forgotten in the migration.
But we can transmit them as distribution models.
We only transmit ones who have a company that is
compatible with the analytic accounts, as there is
a constraint on distribution models.
We also don't transmit defaults who had dates that
are not applicable, to keep the same behaviour
before/after the upgrade.

opw-3167598

closes odoo/upgrade#4351

Related: https://github.com/odoo/odoo/pull/98914
Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>