[FIX] base/16.0: properly get 16.0 views update despite BS5 conversion
Since the bootstrap conversion script made at [1], COW'ed views in user
databases were not able to receive the 16.0 update and were left to
their 15.X version, converted to bootstrap.

What happened was that we were only converting custom/cow'ed views to
bootstrap 5 (the ones that have no external ID), relying on the fact
that those linked to an external ID would be naturally updated to their
16.0 version via the standard module update... this was wrong. Indeed,
we also need to search for views with external ID that have a related
COW'ed view: when updating a generic view after this script, the archs
are compared to know if the related COW'ed views must be updated too or
not. So if we only convert COW'ed views to Bootstrap 5, they won't get
the generic view update as they will be judged different from them (user
customization) because of the BS5 changes that were made.
E.g.
- In 15.0, install website_sale
- Enable eCommerce categories: a COW'ed view is created to enable the
  feature (it leaves the generic disabled and creates an exact copy but
  enabled)
- Migrate to 16.0: you expect your enabled COW'ed view to get the new
  16.0 version of eCommerce categories... but if the COW'ed view was
  migrated to BS5 while the generic was not, they won't be considered
  the same anymore and only the generic view will get the 16.0 update.

Related to the investigation about migration fails of [2].

[1]: https://github.com/odoo/upgrade/commit/cfd58d145b8bbb99b0ec1dcc70bcfbfa3f9b7e20
[2]: https://www.flowup.shop

closes odoo/upgrade#4345

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>