[FIX] account/16.0.1.2: recompute root_id after updating code
The upgrade scripts sanitize the `code` field in `account_account`. In
the related tickets, the code started with a zero-width space. This
character isn't visible, but it does impact the root_id which gets
computed from the code and the hierarchy the account is shown in in the
chart of accounts. If the `root_id` isn't updated together with the
code, it will end up pointing to non-existent rows in the account_root
table (which is a view computed from `account_account`), rendering the
chart of accounts in the UI inaccessible to the user.

In the case of a code '704020', normally the root id is 55048, because
the first character is 7 (ASCII('7') * 1000 = 55000) and the second is
0 (ASCII('0') = 48). 55000 + 48 = 55048.

If there's a zero-width space in front of the string, that will amount
to ASCII('<the zero-width space>') * 1000 = 8203000 for the first
character and ASCII('7') = 55 for the second, leading to a total of
8203055.

How the account_root table is computed in 15.2: https://github
.com/odoo/odoo/blob/73211245425a2e00a376242a8aa7a378322042a7/addons
/account/models/account_account.py#L791 .
How the ORM does the computation in 15.2: https://github
.com/odoo/odoo/blob/73211245425a2e00a376242a8aa7a378322042a7/addons
/account/models/account_account.py#L278 .
The logic is the same in 16.0.

So this fix recomputes the `root_id` on `account_account` following the
sanitizing of `code` to make sure they are in sync.

upg-494960
opw-3184138
opw-3170336

Part-of: odoo/upgrade#4355