[IMP] util/records: new remove_group method
Adds a `remove_group` method to handle group removal.
Usage: through `remove_record` or by directly calling it.

this handles a common issue with foreign keys that have the
`ondelete=restrict` option set. When a group is deleted,
if there are records referencing that group in another table,
an SQL error will occur due to the restriction. To avoid this
error, the `remove_group` method first removes any
dependent records that reference the group, and then proceeds
with deleting the group itself. This ensures that the foreign
key constraint is not violated and the deletion is successful.

upg-483440

Part-of: odoo/upgrade#4357