[FIX] account/16.0.1.2: traceback due to duplicate codes
When sanitizing account codes, every single account was being updated,
which had a performance impact as well as filling the upgrade report
with unhelpful noise. A fix was made to only update necessary accounts:
https://github.com/odoo/upgrade/pull/4378 . This unfortunately breaks
the corner case where a code like `abc-ab` is updated to `abc.ab` while
a code `acb.ab` already exists, resulting in a traceback due to a
constraint violation because the company id + code need to be unique.

For this reason codes that haven't changed shouldn't be filtered out of
the aggregation in `replacement` unconditionally. This fix only filters
out the entire aggregated line if none of the codes inside it changed.

upg-538145
upg-538266
upg-538770
upg-541269

closes odoo/upgrade#4355

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>
Co-authored-by: Alvaro Fuentes <afu@odoo.com>