[IMP] sale_subscription/saas~16.2: move stage to subscription_state
Before this commit, the condition of a subscription sale.order was
defined by thre fields: the state, the stage_id and the
subscription_management. The knowledge of these threee fields was necessary
to classify the subscription in the right category. As these information
were not displayed in all views it was not easy to use the module.

The enterprise PR removes the sale.order.stage model and
subscription_management field to merge them in a single selection field.

closes odoo/upgrade#4386

Enterprise: https://github.com/odoo/enterprise/pull/37536
Related: odoo/enterprise#37536
Signed-off-by: Christophe Simonis <chs@odoo.com>