[IMP] knowledge/saas~16.3: fully implement oeProtected
All Knowledge Behavior anchor should now have the `data-oe-protected="true"`
attribute. This commit is part of enabling the collaborative mode in Knowledge.
Behavior nodes with the `data-prop-name` attribute which should be collaborative
now have the `data-oe-protected="false"` attribute (mutations inside those nodes
will be listened to by the editor once the Component was rendered at least
once).

In order not to make a complex upgrade script to change the value of every
`knowledge.article` record, that attribute is set programmatically in the JS
code during the Behavior component setup.

This upgrade script purpose is to make data (noupdate=1) articles
consistent with this new feature, but it is not mandatory for them to work
properly.

See odoo/odoo#104680
See odoo/enterprise#33483

Task-2821374

closes odoo/upgrade#4384

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>