[IMP] sale_subscription/saas~16.2: sale order logs needs more data
This commit aims to improve the available data on the sale.order.log model.
It will be easier to group by contraction and expansions.

We also chose to save the origin order on the logs to easily know wich is the
parent contract of the order as the origin_order_id is null for the first
contract in a subscription family.

taskid: 3151840
enterprise PR: https://github.com/odoo/enterprise/pull/36390

Part-of: odoo/upgrade#4401