[IMP] base/saas~16.2: remove `search_view` from `ir.actions`
This field has been introduced in
https://github.com/odoo/odoo/commit/c68600cc5397700df4f40095c27723a2b1f90a22

The reason why is not really clear, given the commit message.

But, it looks completely unused nowadays.
We can therefore safely remove it.

In addition, this will save a lot of transferred data to the web client,
as currently all the fields of an action are read when the web client
loads an action,
including this one, which contains a long xml content.

For instance, with just base and the contacts modules installed,
to load the action of the Contacts menu,
it goes from 5.51kB to 1.05kB.

Regarding the use of `search_view` in the Crawler test
of upgrade, it's a mistake: Whether in 10.0 or 16.0,
the web client doesn't use that field,
it asks the `search` views to `get_views`,
except in some circumstances:
when the menu open a dialog form,
then there is no need of a search view.

See odoo/odoo#114041

closes odoo/upgrade#4390

Signed-off-by: Christophe Simonis <chs@odoo.com>