[FIX] account/16.0.1.2: correct root_id for empty code
This is a followup to https://github.com/odoo/upgrade/pull/4355
for the edge case where the code consists entirely out of no longer
accepted characters and becomes an empty string as a result when
sanitized. The current root_id calculation for that (0) is not correct,
which results in a traceback when accessing the chart of accounts. It's
a different one from the traceback in the main PR.

When the groups with root ids are being read for the chart of accounts,
it will filter on the root_id not being `False`:
https://github.com/odoo/odoo/blob/6f90d6924e24e700694111732ee85465f802b22f/addons/web/models/models.py#L332 .
If the root_id is 0 however, the ORM will still convert it to `False`
and it will return a boolean instead of a tuple like for the other valid
values. This will cause the traceback when that tuple is unpacked here:
https://github
.com/odoo/odoo/blob/6f90d6924e24e700694111732ee85465f802b22f/addons/web/models/models.py#L336 .

The reason for that is an inconsistency between the ORM computation for
the root_id field in account_account and the account_root view
definition. The ORM computation checks in Python if the value of the
code evaluates to True. If it doesn't, the root_id gets set to False (so
 NULL in the database). The account root view computation uses the ASCII
computation for each line where the code isn't NULL. This means that for
an empty code it generates an empty line with id 0 that can't be
referenced due to how the ORM computation works. This line is
superfluous and could be removed by filtering out codes that are empty
strings. It doesn't cause any issues on its own, so it's not
particularly high priority.

In the upgrade scripts we're using the inconsistent logic from the
account root query. This fix corrects that to bring it in line with the
ORM logic. Normally empty codes can't happen because the UI won't allow
it (the database does), so it's only in the unlikely case that the
sanitizing of the codes by the upgrade scripts results in an empty code
that the traceback can occur.

Tested with the same modified upg-494960 database as the original PR,
that includes test cases for 0 and 1 digit codes as well as duplicates.

 ORM computation: https://github.com/odoo/odoo/blob/27ef0b93770c4e358a8cf4082afcfeb2da96f6c6/addons/account/models/account_account.py#L321
 Account root view: https://github.com/odoo/odoo/blob/27ef0b93770c4e358a8cf4082afcfeb2da96f6c6/addons/account/models/account_account.py#L974

 upg-494960

closes odoo/upgrade#4392

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>