[FIX] {base,hr_contract_sign}/saas~12.5,base/13.0
Effectively execute the `hr_contract_sign` saas~12.5 script.

Where you call the `force_migration_of_fresh_module` function is
important as it determines from which versions to run the upgrade
scripts.

And why the heck did we deleted a helpdesk field in this script?
As this field is already deleted in `helpdesk_sale/saas~12.4`, it's
probably a copy/paste error.

closes odoo/upgrade#4430

Signed-off-by: Christophe Simonis <chs@odoo.com>