[IMP] project/saas~16.3: Remove commercial_partner_id and project_color
This commit removes 2 fields, commercial_partner_id defined in
`project.project` and `project.task`, and `project_color` defined in
`project.task`.
The reason is that the ``commercial_partner_id`` is not really useful in
project and can be replaced by `partner_id.commercial_partner_id`.
The `project_color` was just a field to colorize the tasks in the same color
of the project in the calendar view.

Community PR: odoo/odoo#115155
Enterprise PR: odoo/enterprise#38142

closes odoo/upgrade#4432

Taskid: 3231143
Signed-off-by: Christophe Simonis <chs@odoo.com>