[FIX] account/16.0: Preserve bank statement attachments
During the migration, the chatter is removed and a new attachment_ids
field is added on account.bank.statement.
Currently, the attachments linked to the chatter are lost.
This commit populate the new attachment_ids field with the existing
attachments during the migration.

This field has been introduced by odoo/odoo#99092

closes odoo/upgrade#4434

Signed-off-by: Christophe Simonis <chs@odoo.com>