[IMP] project/saas~16.3: remove display_project_id and related field
This commit removes the `display_project_id` field in `project.task`
because this field is complicated to be understandable by the users
and so the subtasks feature has been simplify. it also removes
`task_count_with_subtasks` field  in `project.project` because this
field is no longer used.

task-3230063

Community PR: odoo/odoo#115781
Enterprise PR: odoo/enterprise#38378

Part-of: odoo/upgrade#4448