[IMP] account/saas~16.3.1.2: remove tax_audit field
the tax_audit field was introduced in v13 and it served the purpose of
providing users with improved visibility regarding the impact of move
lines on report lines. The computation of the sign for each move line
when resolving tax grid amounts has now been simplified using
the formula (tax_tag_invert * sign of the tag * balance). Consequently,
the tax_audit field has become redundant and is no longer required.

closes odoo/upgrade#4442

Community-pr: odoo/odoo#114616
Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>