[FIX] stock/11.0.1.1: improve performance with hash index
Updating an index many times is a costly operations that greatly
degrades performance. For such workloads a hash index is better. The
problematic queries on this script are of the form
```
UPDATE stock_move_line SET move_id = %s WHERE move_id IN %s
```
that use the index on move_id and at the same time update it. Those
queries are ran millions of times.
```
536031_afu_16.0=> select count(*) from stock_move_line
+---------+
| count   |
|---------|
| 9937500 |
+---------+
```

With the switch to a hash index each individual query went from up to 1
second to almost 0 milliseconds.

Note upg-536031 takes almost 1 day to upgrade to 11.0 and another day to
go from 11.0 to 16.0. After this patch the upgrade to 11.0 last a few
hours.

upg-536031
opw-3105048

closes odoo/upgrade#4447

Signed-off-by: Christophe Simonis <chs@odoo.com>