[FIX] hr_expense/saas~16.3: duplicate journal field
A rework of hr_expense has been introduced in master and landed in 16.2.
https://github.com/odoo/odoo/commit/f79ff3fb374650670e89f231aaf17bb8d6596c25

There was a need to backport it to 16.0 then 16.1.
https://github.com/odoo/odoo/commit/90affb562962d0dc233526e3f15f5b60f383331d

During the backport, a computed fields 'journal_displayed_id'
has been introduced to fix an issue and now needs to be
properly removed in master as it was fixed in 16.2.

closes odoo/upgrade#4443

Community: https://github.com/odoo/odoo/pull/114528
Task-id: 3191597
Related: odoo/odoo#114528
Signed-off-by: Christophe Simonis <chs@odoo.com>