[IMP] account_edi_ubl_cii/saas~16.2: select format based on partner
Add a selection field on the partner to choose the EDI format. Merge the
cii_ and ubl_ fields.

Add an endpoint_value and eas_code on the partner. These values are read
when generating the Bis 3 (or one of its derivatives) to fill the
EndpointID and the @schemeID in the xml.

We can remove l10n_nl_oin and l10n_nl_kvk which are only used by
electronic invoicing, and replace every usage of them by the new fields
endpoint_value and eas_code (since for NL partners, they should
necessarily represent the KVK or OIN number: eas code '0106' or '0190').

We can also remove the l10n_lu_peppol_id module which was only there to
allow an arbitrary number to be set on the EndpointID (for instance, for
public administration that doesn't have a vat number).

task-3232845

closes odoo/upgrade#4445

Signed-off-by: Christophe Simonis <chs@odoo.com>