[FIX] base/16.0: ignore o2m check for manual fields
Due to the change introduced in odoo/upgrade@0906e31c we need to ignore
fields that may be loaded after the current one2many manuel field is
updated. To avoid issues like
```
  File "/tmp/tmpwy7esdlr/migrations/base/0.0.0/pre-models-update-custom-fields.py", line 17, in _auto_init
    super(Base, self.with_context(update_custom_fields=True))._auto_init()
  File "/home/odoo/src/odoo/16.0/odoo/models.py", line 2574, in _auto_init
    new = field.update_db(self, columns)
  File "/home/odoo/src/odoo/16.0/odoo/fields.py", line 4314, in update_db
    raise UserError(_("No inverse field %r found for %r") % (self.inverse_name, self.comodel_name))
odoo.exceptions.UserError: No inverse field 'x_studio_factura_relacionada' found for 'stock.picking'
```
due to a manual field that would be loaded later.

Also the change in odoo/upgrade@0906e31c is only needed while upgrading
to 16.0 because we for the update_db of manual fields to ensure the
translation are stored as jsonb column.

tbg-178
upg-596381

closes odoo/upgrade#4444

Signed-off-by: Christophe Simonis <chs@odoo.com>