[FIX] stock/16.0: enable create access for location
IN Odoo 16 , groups_id field has been removed from ir.ui.view.
The `create` button should not be visible in the tree and form views
of stock locations when the multi locations feature is not enabled.

As we would like to remove the `groups_id` feature of the `ir.ui.view`,
we must find an alternative.
This is the only case using `groups_id` to remove the create/edit
button, so it's not like it's such a common way
to achieve this goal.

so if we have multiple-location group enabled then
just de-activate that group. so create button become
visible for location and user with multi-location
group are able to create location.

see: https://github.com/odoo/odoo/commit/8785bc74770f1029aef16582c8bcf8aa63d90b31

upg-536718
opw-3199576

closes odoo/upgrade#4450

Signed-off-by: Christophe Simonis <chs@odoo.com>