[FIX] sale/16.0: ensure references exist
The mail template `sale.mail_template_sale_confirmation` uses some
external references that are not yet loaded when the template is
updated. Since the original version of the template in the DB was
noupdate and didn't use the new references there is no error, but when
we force the update from xml the reference cannot be found.

closes odoo/upgrade#4461

Signed-off-by: Christophe Simonis <chs@odoo.com>