[IMP] sale_planning/saas~16.3: remove deleted view
Previously, there was a button "Plan Orders" in `planning` gantt view,
that only appeared when `sale_planning` was installed,
because it was used to create shifts out of the confirmed sales orders
that contained products of type `service` with `planning_enabled`.

This button's label is now "Auto Plan". It keeps this functionality,
but is also used to assign resources on open (unassigned) shifts.
Therefore, we need it in `planning` already,
and so we don't need to override `planning` gantt view in `sale_planning`.

Also, we created a field `Resource.default_planning_role_id`
that has the same value as `Resource.default_planning_role_id`,
and the latter is now related to the former.
This commit does that, and removes the latter's column.

Related PR: https://github.com/odoo/enterprise/pull/34581
task-2959882

closes odoo/upgrade#4465

Related: odoo/enterprise#34581
Signed-off-by: Christophe Simonis <chs@odoo.com>