[FIX] analytic/16.0.1.1,account/16.0.1.2: transmit empty tag info
In the migration to 16, the analytic tags have been removed.
If they didn't have any distribution, the information is lost.
To help users with this, it has been decided that we should create
analytic accounts from the tag (with the name from the tag and
belonging to a new plan 'Former Tag').
With this, it will appear in the distribution.
To stay consistent, analytic lines have been created from these
new accounts.

opw-3208917

closes odoo/upgrade#4479

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>