[FIX] base/tests: mute some logger
Mute the `odoo.upgrade.helpers` logger when testing the `merge` strategy
on xmlid rename.

The `sale_subscription` upgrade keeps the `sale_subscription_line` table
when removing the model. This table has an FK on `res_currency`, when
merging two currencies, a warning is shown because the model of the
`sale_subscription_line` table cannot be found (and that's expected).

Avoid this by explicitly muting the logger.

closes odoo/upgrade#4484

Signed-off-by: Christophe Simonis <chs@odoo.com>