[FIX] website/saas~13.5: handling of xml/html for parallax conversion
Commit ff86c23 adds parsing of html data for the "parallax" website
snippet conversion. The issue arises from the fact that while parsed
as html the lxml tree is then serialized as xml.
One key difference between the html/xml parsers is that `html`
preserves CRLF EOL, while `xml` strips them at parsing.
If a `\r` (CR) is encountered in the etree when serializing as `xml`,
it will get entity-escaped to `&#13;`.
This pollutes the resulting xml and makes it unparseable if such
character appears after the last tag ("extra content" error).

To be properly converted, xml and html must be parsed and serialized
with the correct parser/method, these cannot be safely mixed.

closes odoo/upgrade#4488

Signed-off-by: Christophe Simonis <chs@odoo.com>