[FIX] base/saas~16.0: force upgrade of l10n_mx_edi module
Force upgrade of the l10n_mx_edi module because in the database,
a recursion error (RecursionError: maximum recursion depth exceeded) is thrown
when the l10n_mx_edi module is loading with a large number of records.

When in the production database of the customer, l10n_mx_reports is installed
and it contains a large number of records in the database. So in that case,
a recursion error occurs.

In version 15 the current state of the l10n_mx_edi module is uninstall and
during migration, the state of that module is installed because
it depends on l10n_mx_reports and the auto-install of that module is true.

closes odoo/upgrade#4492

Signed-off-by: Christophe Simonis <chs@odoo.com>