[IMP] base_vat/saas~16.3: Remove VIES message field
See: https://github.com/odoo/odoo/pull/116391

The above PR adapts the existing VIES VAT check mechanism, used to test
the validity of a VAT code against the VIES EU database. The commit
introduces a stored boolean field, visible from the partner form, called
"vies_valid", which replaces a computed, non-stored 'message' field,
which was used to display a banner indicating the VAT had failed the
VIES check on the partner form. The stored field is needed as it is used
to determine the applicability of some fiscal positions.

This commit removes the computed non-stored 'message' field and adds the
column "vies_valid" on the partner.

A query is executed in order to update the value of the vies_check field
on the partner (which is false by default) to true, depending on the
partner's usage in the past. This is necessary because the vies_valid
field affects the way that fiscal positions are applied, and the user
should not have to input the value for all the partners on their db.
Thus, the field should automatically be set to true for partners that
have invoices that appear to have an intra-community context.

When a partner has:
     - A vat, looking as though it belongs to a European nation
     - invoices, of which the most recent invoice looks as though
       it has been used in an intra-community transaction because:
          a) it has a fiscal position that looks to be an
          intra-community fiscal position
          b) it is between a European company, and partner and has no
          total tax
Then the vies_valid field is set to true, otherwise it remains false.

closes odoo/upgrade#4498

Task-id: `3218194`
Related: odoo/odoo#116391
Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>