[FIX] base/0.0.0: correctly build views' tree
Before this patch the way we checked if there was a custom view in the
inheritance tree was wrong. For cases like:
```
 P0
  | \
  | E1 -> custom view
 P1
```
where P/E means primary/extension respectively. We didn't take into
account the extension E1 in the tree of P1. That is when we called
`custom_view_in_tree(P1)` since it was already a primary view we didn't
check E1.

The effect this had is that when P1 is a standard view the call to
```
validate_view(view, True)  # dummy validators
```
was not done. Thus we never got a chance to catch this error in the
fixer. This omission was not that bad for two reasons: it only impacted
standard views -- that we never fix, just restore; and the type of
errors that it didn't uncover are currently not fixed by our code -- it
requires disabling custom views in the whole tree to check which one is
breaking the standard view.

`_get_combined_arch` in standard code filters out some views according
to `load_all_views` value when checking the XML structure. Here are the
relevant pointers.
https://github.com/odoo/odoo/blob/1a4a70e9fae83700415bb0da51a7063bf889d751/odoo/addons/base/models/ir_ui_view.py#L993
https://github.com/odoo/odoo/blob/1a4a70e9fae83700415bb0da51a7063bf889d751/odoo/addons/base/models/ir_ui_view.py#L688-L689

The new implementation follows standard code in `_get_inheriting_views`
Namely we recover the path of ancestors `P1, P0` in the example, then
add all _extension_ descendants. See:
https://github.com/odoo/odoo/blob/1a4a70e9fae83700415bb0da51a7063bf889d751/odoo/addons/base/models/ir_ui_view.py#L650-L668

upg-649374 is an example of the issue: the crawler fails but there is no
error in the checks done by the fixer (for view with id=2096). After
this patch the view produces an error (see below). Note that the view is
standard although we call it _custom_ (in some sense it is "custom") in
the logs.
```
2023-03-30 22:26:43,101 22 ERROR 649374_afu_16.0 odoo.addons.base.maintenance.migrations.base.pre-models-ir_ui_view: Invalid custom view sale_temporal.product_template_tree_view for model product.template
Traceback (most recent call last):
  File "/tmp/tmp7ep5k9n0/migrations/base/0.0.0/pre-models-ir_ui_view.py", line 713, in validate_view
    view._check_xml()
  File "/tmp/tmp7ep5k9n0/migrations/base/0.0.0/pre-models-ir_ui_view.py", line 429, in _check_xml
    super(IrUiView, record)._check_xml()
  File "/home/odoo/src/odoo/16.0/odoo/addons/base/models/ir_ui_view.py", line 434, in _check_xml
    raise err from None
  File "/home/odoo/src/odoo/16.0/odoo/addons/base/models/ir_ui_view.py", line 425, in _check_xml
    combined_arch = view._get_combined_arch()
  File "/home/odoo/src/odoo/16.0/odoo/addons/base/models/ir_ui_view.py", line 991, in _get_combined_arch
    arch = root.with_prefetch(tree_views._prefetch_ids)._combine(hierarchy)
  File "/home/odoo/src/odoo/16.0/odoo/addons/base/models/ir_ui_view.py", line 932, in _combine
    combined_arch = view.apply_inheritance_specs(combined_arch, arch)
  File "/home/odoo/src/enterprise/16.0/web_studio/models/ir_ui_view.py", line 592, in apply_inheritance_specs
    return super(View, self).apply_inheritance_specs(source, specs_tree,
  File "/home/odoo/src/odoo/16.0/odoo/addons/base/models/ir_ui_view.py", line 869, in apply_inheritance_specs
    self._raise_view_error(str(e), specs_tree)
  File "/home/odoo/src/odoo/16.0/odoo/addons/base/models/ir_ui_view.py", line 746, in _raise_view_error
    raise err from from_exception
odoo.exceptions.ValidationError: Ocurrió un error al validar la vista:

Element '<field name="standard_price">' cannot be located in parent view
```

For the sake of completeness. The issue in upg-649374 is that, following
the example notation, E1 removes an anchor in P0 used by P1, both P0 and
P1 are standard views. Only when validating P1 with all validators
(dummy ones in the comment) the errors happens in the fixer. Note that
P0 is correctly validated, as well as E1. It's only P1 _with_ E1 applied
to P0 that is invalid.

Part-of: odoo/upgrade#4507